% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_fit.R
\name{rf_fit}
\alias{rf_fit}
\title{This function takes a data frame of summary course enrollments
and returns a Random Forest Model object for the purpose of
predicting total Course Enrollments for a given Subject Area
(e.g. Graduate Data Science courses)}
\usage{
rf_fit(course_set)
}
\arguments{
\item{course_set}{A data frame containing the required 7 columns}
}
\value{
random forest model object
}
\description{
This function takes a data frame of summary course enrollments
and returns a Random Forest Model object for the purpose of
predicting total Course Enrollments for a given Subject Area
(e.g. Graduate Data Science courses)
}
\examples{
rf_fit(course_set)
}
