% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecoursedata.R
\docType{data}
\name{samplecoursedata}
\alias{samplecoursedata}
\title{Sample Course Data}
\format{A data frame with 24390 rows and 7 variables:
\describe{
  \item{Term}{Course Schedule/Enrollment Term}
  \item{CourseID}{Unique Course Identifier}
  \item{Subject}{Subject Area of the Course}
  \item{MaxUnits}{Maximum Unit Value for the Course}
  \item{TotalCapacity}{Total Combined Class Section Capacity}
  \item{EnrlTotal}{Total Combined Class Section Enrollment}
  \item{TotalSections}{Total Scheduled Sections for the Course}
  }}
\usage{
samplecoursedata
}
\description{
Combined Course Enrollment Statistics from Fall 2016 thru Summer 2018
}
\keyword{datasets}
