% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score Function (Predict Total Course Enrollment W/ 4 Inputs)}
\usage{
score(courseID, maxunits, totalcapacity, totalsections)
}
\arguments{
\item{CourseID}{6 digit integer which is the unique course identifier}

\item{MaxUnits}{2 digit maximum integer}

\item{TotalCapacity}{3 digit maximum integer}

\item{TotalSections}{3 digit maximum integer}
}
\examples{
  score(125768,3,75,4)
  score(154233,3,63,4)
}
