% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_linear.R
\name{score_linear}
\alias{score_linear}
\title{Score Linear Function - Predict Total Course Enrollment using Linear Regression (4 Parameters)}
\usage{
score_linear(courseID, maxunits, totalcapacity, totalsections)
}
\arguments{
\item{CourseID}{6 digit integer which is the unique UC Berkeley course identifier}

\item{MaxUnits}{2 digit(max) integer}

\item{TotalCapacity}{3 digit(max) integer}

\item{TotalSections}{3 digit(max) integer}
}
\examples{
  score(125768,3,75,4)-DATASCI W266
  score(154233,3,63,4)-DATASCI W200
  score(114496,4,107,4)-MATH 1A
  score(114538,4,36,1)-MATH 141
  score(118027,4,100,1)-PHILOS 12A
}
